modernize flaky request hooks test and remove a redundant one
